### 2024-08-26 d62d15e
Benchmark results, update to toml, and documentation
