+++ benchmark/spe11a/result3.txt
@@ -1,0 +1,54 @@
+"""Set the full path to the flow executable and flags"""
+mpirun -np 32 flow --partition-method=2 --edge-weights-method=2 --imbalance-tol=1.1 --min-strict-cnv-iter=5 --tolerance-cnv-relaxed=1e-2 --tolerance-cnv=1e-3 --tolerance-mb=1e-7 --linear-solver=cprw --newton-min-iterations=1 --enable-tuning=true --enable-opm-rst-file=true --output-extra-convergence-info=steps,iterations --enable-drift-compensation=0
+
+"""Set the model parameters"""
+spe11a master     #Name of the spe case (spe11a, spe11b, or spe11c) and OPM Flow version (master or release)
+complete gaswater #Name of the co2 model (immiscible or complete) and co2store implementation (gaswater or gasoil [oil properties are set to water internally in OPM flow])
+corner-point      #Type of grid (cartesian, tensor, or corner-point)
+2.8 0.01 1.2      #Length, width, and depth [m]
+280               #If cartesian, number of x cells [-]; otherwise, variable array of x-refinment
+1                 #If cartesian, number of y cells [-]; otherwise, variable array of y-refinment [-] (for spe11c)
+20,7,1,5,6,3,8,8,12,8,10,10,6,16,6,28,28,1  #If cartesian, number of z cells [-]; if tensor, variable array of z-refinment; if corner-point, fix array of z-refinment (18 entries)
+20 20             #Temperature bottom and top rig [C]
+1.2 110000 1      #Datum [m], pressure at the datum [Pa], and multiplier for the permeability in the z direction [-] 
+1e-9 1.6e-5       #Diffusion (in liquid and gas) [m^2/s]
+0 0               #Rock specific heat and density (for spe11b/c)
+0 0 0             #Added pore volume on top boundary (for spe11a [if 0, free flow bc]), pore volume on lateral boundaries, and width of buffer cell [m] (for spe11b/c)
+0 0               #Elevation of the parabola and back boundary [m] (for spe11c)
+
+"""Set the saturation functions"""
+(max(0, (s_w - swi) / (1 - swi))) ** 2                                                        #Wetting rel perm saturation function [-]
+(max(0, (1 - s_w - sni) / (1 - sni))) ** 2                                                    #Non-wetting rel perm saturation function [-]
+penmax * math.erf(pen * ((s_w-swi) / (1.-swi)) ** (-(1.0 / 2)) * math.pi**0.5 / (penmax * 2)) #Capillary pressure saturation function [Pa]
+(np.exp(np.flip(np.linspace(0, 5.0, npoints))) - 1) / (np.exp(5.0) - 1)                       #Points to evaluate the saturation functions (s_w) [-]
+
+"""Properties sat functions"""
+"""swi [-], sni [-], pen [Pa], penmax [Pa], npoints [-]"""
+SWI1 0.32 SNI1 0.1 PEN1 1500 PENMAX1 2500 NPOINTS1 1000
+SWI2 0.14 SNI2 0.1 PEN2  300 PENMAX2 2500 NPOINTS2 1000
+SWI3 0.12 SNI3 0.1 PEN3  100 PENMAX3 2500 NPOINTS3 1000
+SWI4 0.12 SNI4 0.1 PEN4   25 PENMAX4 2500 NPOINTS4 1000
+SWI5 0.12 SNI5 0.1 PEN5   10 PENMAX5 2500 NPOINTS5 1000
[…]
