### 2024-08-26 ef0ab90
Option for aditional tuning parameters in the deck
