+++ requirements.txt
@@ -1,8 +1,0 @@
-mako
-matplotlib
-pandas
-Pyarrow
-resdata
-rtree
-scipy
-shapely
