### 2024-08-27 c63c4ee
Documentation for the convective model
