+++ src/pyopmspe11/utils/runs.py
@@ -8,22 +8,28 @@ import os
 import subprocess
 
 
-def simulations(dic, deck, folder):
+def simulations(dic, deck, folder, dryrun):
     """
     Run OPM Flow
 
     Args:
         dic (dict): Global dictionary\n
         deck (str): Name of the input deck\n
-        folder (str): Name of destination of the output files
+        folder (str): Name of destination of the output files\n
+        dryrun (bool): Run a dry simulation for the initial files.
 
     Returns:
         None
 
     """
+    flag = ""
+    flow = dic["flow"]
+    if dryrun:
+        flow = dic["only_flow"]
+        flag = "--enable-dry-run=1"
     os.system(
-        f"{dic['flow']} --output-dir={dic['exe']}/{dic['fol']}/{folder} "
-        f"{dic['exe']}/{dic['fol']}/deck/{deck}.DATA & wait\n"
+        f"{flow} --output-dir={dic['exe']}/{dic['fol']}/{folder} "
+        f"{dic['exe']}/{dic['fol']}/deck/{deck}.DATA {flag} & wait\n"
     )
 
 
