+++ examples/finner_grids/spe11c_cart_168_100_120.txt
@@ -3,7 +3,7 @@ mpirun -np 33 flow  --zoltan-imbalance-tol=1.05 --tolerance-mb=1e-7 --linear-sol
 
 """Set the model parameters"""
 spe11c master     #Name of the spe case (spe11a, spe11b, or spe11c) and OPM Flow version (master or release)
-complete gaswater #Name of the co2 model (immiscible or complete) and co2store implementation (gaswater or gasoil [oil properties are set to water internally in OPM flow])
+complete gaswater #Name of the co2 model (immiscible, convective [convective requires a Flow version newer than 22-08-2024], or complete) and co2store implementation (gaswater or gasoil [oil properties are set to water internally in OPM flow])
 cartesian         #Type of grid (cartesian, tensor, or corner-point)
 8400 5000 1200    #Length, width, and depth [m]
 168               #If cartesian, number of x cells [-]; otherwise, variable array of x-refinment
