+++ examples/hello_world/spe11a.txt
@@ -3,7 +3,7 @@ flow --tolerance-mb=1e-7 --linear-solver=cprw --enable-tuning=true --newton-min-
 
 """Set the model parameters"""
 spe11a release    #Name of the spe case (spe11a, spe11b, or spe11c) and OPM Flow version (master or release)
-complete gaswater #Name of the co2 model (immiscible or complete) and co2store implementation (gaswater or gasoil [oil properties are set to water internally in OPM flow])
+complete gaswater #Name of the co2 model (immiscible, convective [convective requires a Flow version newer than 22-08-2024], or complete) and co2store implementation (gaswater or gasoil [oil properties are set to water internally in OPM flow])
 tensor            #Type of grid (cartesian, tensor, or corner-point)
 2.8 0.01 1.2      #Length, width, and depth [m]
 4,15,10,10,9,7    #If cartesian, number of x cells [-]; otherwise, variable array of x-refinment
