+++ examples/finner_grids/spe11b_cart_1m_nldd_tuned.txt
@@ -3,7 +3,7 @@ mpirun -np 12 flow --nonlinear-solver=nldd --matrix-add-well-contributions=1 --t
 
 """Set the model parameters"""
 spe11b master     #Name of the spe case (spe11a, spe11b, or spe11c) and OPM Flow version (master or release)
-complete gaswater #Name of the co2 model (immiscible or complete) and co2store implementation (gaswater or gasoil [oil properties are set to water internally in OPM flow])
+complete gaswater #Name of the co2 model (immiscible, convective [convective requires a Flow version newer than 22-08-2024], or complete) and co2store implementation (gaswater or gasoil [oil properties are set to water internally in OPM flow])
 cartesian         #Type of grid (cartesian, tensor, or corner-point)
 8400 1 1200       #Length, width, and depth [m]
 8400              #If cartesian, number of x cells [-]; otherwise, variable array of x-refinment
