+++ examples/finner_grids/spe11b_cp_10mish.txt
@@ -3,7 +3,7 @@ mpirun -np 33 flow --tolerance-mb=1e-7 --linear-solver=cprw --newton-min-iterati
 
 """Set the model parameters"""
 spe11b master     #Name of the spe case (spe11a, spe11b, or spe11c) and OPM Flow version (master or release)
-complete gaswater #Name of the co2 model (immiscible or complete) and co2store implementation (gaswater or gasoil [oil properties are set to water internally in OPM flow])
+complete gaswater #Name of the co2 model (immiscible, convective [convective requires a Flow version newer than 22-08-2024], or complete) and co2store implementation (gaswater or gasoil [oil properties are set to water internally in OPM flow])
 corner-point      #Type of grid (cartesian, tensor, or corner-point)
 8400 1 1200       #Length, width, and depth [m]
 840               #If cartesian, number of x cells [-]; otherwise, variable array of x-refinment
