+++ benchmark/spe11c/r4_cp_8m-8mish-8mish.txt
@@ -1,20 +1,20 @@
-"""Set the full path to the flow executable and flags"""
-mpirun -np 64 flow --relaxed-max-pv-fraction=0 --partition-method=2 --edge-weights-method=2 --imbalance-tol=1.1 --tolerance-mb=1e-7 --enable-tuning=true --enable-opm-rst-file=true --output-extra-convergence-info=steps,iterations --newton-min-iterations=1 --linear-solver=cpr_trueimpes --time-step-control=newtoniterationcount --newton-max-iterations=13 --solver-growth-factor=1.9 --time-step-control-growth-rate=1.5 --enable-drift-compensation=0
+"""Set the full path to the flow executable and flags""" 
+mpirun -np 1024 flow --partition-method=2 --edge-weights-method=2 --imbalance-tol=1.1 --tolerance-mb=1e-7 --linear-solver=cprw --enable-tuning=true --enable-opm-rst-file=true --output-extra-convergence-info=steps,iterations --newton-min-iterations=1
 
 """Set the model parameters"""
-spe11b master     #Name of the spe case (spe11a, spe11b, or spe11c) and OPM Flow version (master or release)
+spe11c master     #Name of the spe case (spe11a, spe11b, or spe11c) and OPM Flow version (master or release)
 complete gaswater #Name of the co2 model (immiscible, convective [convective requires a Flow version newer than 22-08-2024], or complete) and co2store implementation (gaswater or gasoil [oil properties are set to water internally in OPM flow])
-cartesian         #Type of grid (cartesian, tensor, or corner-point)
-8400 1 1200       #Length, width, and depth [m]
-8400              #If cartesian, number of x cells [-]; otherwise, variable array of x-refinment
-1                 #If cartesian, number of y cells [-]; otherwise, variable array of y-refinment [-] (for spe11c)
-1200              #If cartesian, number of z cells [-]; if tensor, variable array of z-refinment; if corner-point, fix array of z-refinment (18 entries)
-70 40             #Temperature bottom and top rig [C]            
+corner-point      #Type of grid (cartesian, tensor, or corner-point)
+8400 5000 1200    #Length, width, and depth [m]
+1050              #If cartesian, number of x cells [-]; otherwise, variable array of x-refinment
+118,122,125,122,118 #If cartesian, number of y cells [-]; otherwise, variable array of y-refinment [-] (for spe11c)
+9,7,1,5,5,3,7,7,10,7,7,8,6,15,5,24,24,7 #If cartesian, number of z cells [-]; if tensor, variable array of z-refinment; if corner-point, fix array of z-refinment (18 entries)
+70 36.12          #Temperature bottom and top rig [C]            
 300 3e7 0.1       #Datum [m], pressure at the datum [Pa], and multiplier for the permeability in the z direction [-] 
 1e-9 2e-8         #Diffusion (in liquid and gas) [m^2/s]
 8.5e-1 2500       #Rock specific heat and density (for spe11b/c)
 0 5e4 1           #Added pore volume on top boundary (for spe11a [if 0, free flow bc]), pore volume on lateral boundaries, and width of buffer cell [m] (for spe11b/c)
-0 0               #Elevation of the parabola and back boundary [m] (for spe11c)
+150 10            #Elevation of the parabola and back [m] (for spe11c)
 
 """Set the saturation functions"""
[…]

[…]
