### 2024-09-13 #7043 @rwegener2
👋🏻 @rowanc1, would you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
