### 2024-10-01 5f4cfaa
spe11a 1mm case with stricter tolerances
