### 2024-10-15 #7357 @editorialbot
[REVIEW]: pyopmspe11: A Python framework using OPM Flow for the SPE11 benchmark project
**Submitting author:** @daavid00 (<a href="http://orcid.org/0000-0002-3343-1005">David Landa-Marbán</a>)
**Repository:** https://github.com/OPM/pyopmspe11
**Branch with paper.md** (empty if default branch):
**Version:** v2024.10
**Archive:** 10.6084/m9.figshare.28284845.v1
