### 2024-10-16 #7357 @daavid00
> 👋🏼 @daavid00 @MatthewFlamm, @gassmoeller this is the review thread for the paper. All of our communications will happen here from now on.

Many thanks @rwegener2 for following this and many thanks @MatthewFlamm and @gassmoeller for accepting reviewing it; @totto82 and me we look forward to the reviewing process.
