### 2024-11-14 OPM/pyopmspe11#83 @gassmoeller
Expand Contributing.md to invite contributions
The current [contributing.md](https://github.com/OPM/pyopmspe11/blob/main/CONTRIBUTING.md) is a bit short. It sets some guidelines for contributions and discusses the process for users already familiar with contributing to github projects, but little else. For a small project like this a 10 page Contributing.md is maybe too much, but a bit of additional information would be good. E.g. the sentence `This project welcomes and values your contributions.` means a lot for potential contributors. Since most of this test is templateable there are good examples out there, e.g. maybe you can take over some of the text from https://contributing.md/example/.

Related to https://github.com/openjournals/joss-reviews/issues/7357.
