### 2024-11-14 OPM/pyopmspe11#83 @gassmoeller
Oh I just found the following in README.md: `Contributions are more than welcome using the fork and pull request approach.`  which is a good start, but should be repeated in the Contributing.md. Still the text of Contributing.md should be extended.

Just to clarify, the JOSS criteria for contributing statements are: Are there clear guidelines for third parties wishing to 1. Contribute to the software 2. Report issues or problems with the software 3. Seek support
