### 2024-11-15 #7357 @gassmoeller
I finished my review, but I would like to wait with a recommendation (major/minor revisions) until the authors have commented on my raised issues. In particular I would like to see how the discussion in https://github.com/OPM/pyopmspe11/issues/85 about my largest concern (the chosen input file format) develops. Generally I think the software is a useful addition and fulfills most of the JOSS publication criteria already. I will report back when the authors have commented on my issues.
