### 2024-11-15 OPM/pyopmspe11#86 @MatthewFlamm
Versioning with tags
This project would be more usable for downstream users if there are versions with tags and/or releases.  Currently installation is done through a commit ref on the git repository, either local or remote.  This makes it hard to do version control for downstream users.  Most users will want to use a versioned release rather than the bleeding edge, which is typically less stable.

Right now the version is hard coded here for the package

https://github.com/OPM/pyopmspe11/blob/17233487c19ced039476e5da427a4b93b72aaf8a/pyproject.toml#L12

and many commits have the same version.

Related to https://github.com/openjournals/joss-reviews/issues/7357
