### 2024-11-15 OPM/pyopmspe11#86 @MatthewFlamm
An example, is to use [setuptools_scm](https://setuptools-scm.readthedocs.io/en/stable/), which will auto-version your package for each commit based on whether there is a tag or not.  I want to make explicit here that there are many ways to do this, `setuptools_scm` is just one such way.
