### 2024-11-15 OPM/pyopmspe11#87 @MatthewFlamm
Releases to package repositories
It would be easier to use for users if they could `pip install pyopmspe11` from the PyPI repository.  This also makes it easier to get a specific released version compared to the git based options [related to #86].  Other options to consider are `conda` (related to #79) through `conda-forge`.

Related to https://github.com/openjournals/joss-reviews/issues/7357
