### 2024-11-19 OPM/pyopmspe11#88 @MatthewFlamm
Running code from visualization module requires `LaTeX`
It isn't documented that `LaTeX` is a required dependency, however when attempting to run the example code `hello_world` with `pyopmspe11 -i spe11b.txt -o spe11b -m all -g all -t 5 -r 50,1,15 -w 1`, it fails when it cannot detect these packages.  This line seems to enforce this requirement

https://github.com/OPM/pyopmspe11/blob/17233487c19ced039476e5da427a4b93b72aaf8a/src/pyopmspe11/visualization/plotting.py#L21

In my opinion, it would be better to remove this requirement, if possible without removing key functionality, as it will make it more widely usable.  Otherwise, it should be documented as a requirement in some way.

Related to https://github.com/openjournals/joss-reviews/issues/7357.
