### 2024-11-19 OPM/pyopmspe11#89 @MatthewFlamm
Expanding references in paper
This is not my specific area of expertise, but the [Related](https://opm.github.io/pyopmspe11/related.html) documentation section has several related packages, and some are also using OPM Flow. These are not referenced in the paper.md.  I would recommend briefly commenting on the how this work is different from the related packages, particularly the ones related to OPM Flow and/or CO2 storage.

Related to https://github.com/openjournals/joss-reviews/issues/7357
