### 2024-11-19 OPM/pyopmspe11#82 @MatthewFlamm
I ran into the first error as well: `NameError: name 'OpmRestart' is not defined`
