### 2024-11-19 #7357 @MatthewFlamm
> In particular I would like to see how the discussion in [OPM/pyopmspe11#85](https://github.com/OPM/pyopmspe11/issues/85) about my largest concern (the chosen input file format) develops.

While I think that the points in my comment above are all important to address to approve, they will only be low/moderate amounts of work IMO.  [OPM/pyopmspe11#85](https://github.com/OPM/pyopmspe11/issues/85) raised by the other reviewer will require the most significant work to address, but the suggestion is a very good one.  Since this package is mainly designed to provide a standardized methodology *via configuration file* to run code against a benchmark, it is more important that this part of the package conforms to best practice.  I think that discussion will be an important one for the requirements for "Substantial scholarly effort" (in the same sense as the above comment) and for "Functionality".
