### 2024-11-20 OPM/pyopmspe11#92 @MatthewFlamm
Provide optional dependencies for opm
The documentation suggests optionally installing `opm` python package using `pip install opm`

> For not macOS users, to install the Python OPM package, execute in the terminal pip install opm.

It would be convenient for users to install everything in one go using `pip install pyopmspe11[opm]`  by declaring it as an optional dependency in pyproject.toml.  https://packaging.python.org/en/latest/guides/writing-pyproject-toml/#dependencies-and-requirements.

While nice for users, this is not directly relevant to the review in https://github.com/openjournals/joss-reviews/issues/7357.
