### 2024-12-02 767d5a1
Option to hide Python warnings
