### 2024-12-02 3ae93ef
Make possible to run tests independly
