### 2024-12-02 6f99fe1
Expand Contributing.md to invite collaboration and run tests
