+++ README.md
@@ -12,22 +12,24 @@ Here we use the [_OPM-Flow_](https://opm-project.org/?page_id=19) simulator.
 
 ## Installation
 You will first need to install
-* Flow (https://opm-project.org, Release 2024.04 or current master branches)
+* Flow (https://opm-project.org, Release 2024.10 or current master branches)
 
-To install the _pyopmspe11_ executable in an existing Python environment: 
+To install the _pyopmspe11_ executable from the development version: 
 
 ```bash
 pip install git+https://github.com/OPM/pyopmspe11.git
 ```
 
-If you are interested in modifying the source code, then you can clone the repository and install the Python requirements in a virtual environment with the following commands:
+If you are interested in a specific version (e.g., v2024.04) or in modifying the source code, then you can clone the repository and install the Python requirements in a virtual environment with the following commands:
 
 ```bash
 # Clone the repo
 git clone https://github.com/OPM/pyopmspe11.git
 # Get inside the folder
 cd pyopmspe11
-# Create virtual environment
+# For a specific version (e.g., v2024.04), or skip this step (i.e., edge version)
+git checkout v2024.04
+# Create virtual environment (to specific Python, python3.12 -m venv vpyopmspe11)
 python3 -m venv vpyopmspe11
 # Activate virtual environment
 source vpyopmspe11/bin/activate
@@ -39,7 +41,7 @@ pip install -e .
 pip install -r dev-requirements.txt
 ``` 
 
-See the [_installation_](https://OPM.github.io/pyopmspe11/installation.html) for further details on building OPM Flow from the master branches in Linux, Windows, and macOS, as well as the OPM Python package.
+See the [_installation_](https://OPM.github.io/pyopmspe11/installation.html) for further details on building OPM Flow from the master branches in Linux, Windows, and macOS, as well as the OPM Python package and LaTeX dependencies.
 
 ## Running pyopmspe11
 You can run _pyopmspe11_ as a single command line:
@@ -57,3 +59,4 @@ The pyopmspe11 package is being funded by the [_HPC Simulation Software for the 
 [project no. 331841].
 This is work in progress. [_Here_](https://www.spe.org/en/csp/) is the link to the SPE11 details.
 Contributions are more than welcome using the fork and pull request approach.
+For a new feature, please request this by raising an issue.
