+++ docs/_sources/about.rst.txt
@@ -7,5 +7,9 @@ About pyopmspe11
 
 This **pyopmspe11** package is being funded by the `HPC Simulation Software for the Gigatonne Storage Challenge project <https://www.norceresearch.no/en/projects/hpc-simulation-software-for-the-gigatonne-storage-challenge>`_ 
 [project number 622059] and `Center for Sustainable Subsurface Resources (CSSR) <https://cssr.no>`_ [project no. 331841].
-This is work in progress. `Here <https://www.spe.org/en/csp/>`_ is the link to the spe11 details.
+
+`Here <https://www.spe.org/en/csp/>`_ is the link to the spe11 details.
+
 Contributions are more than welcome using the fork and pull request approach.
+
+For a new feature, please request this by raising an issue.
