### 2024-12-02 bfd357b
Removing unecessary text
