### 2024-12-03 d169ed5
Paper revised version
