+++ paper/paper.md
@@ -23,49 +23,31 @@ bibliography: paper.bib
 
 # Summary
 
-The imperative to achieve climate change goals and the increasing worldwide demand for energy 
-have made carbon capture and storage (CCS) technology more relevant today.
-Since utilizing computational models is essential for planning large-scale CCS projects, then it is crucial
-to benchmark simulation tools to enhance confidence in the results. Inspired by
-a recent validation study for laboratory-scale CO$_2$ storage [@Flemisch:2024], a new comparative
-solution project (CSP) was launched to simulate both lab- and field-scale CO$_2$ storage [@Nordbotten:2024].
-As part of the Early Access team, the authors have developed and made open the
-`pyopmspe11` tool which facilitates reproducible solutions to the benchmark. This tool serves as a 
-common starting point for developing and testing new simulation technology. It is expected that its impact 
-will extend far beyond the initial benchmark study.
+The imperative to achieve climate change goals and the increasing worldwide demand for energy have made carbon capture and storage (CCS) technology more relevant today. Since utilizing computational models is essential for planning large-scale CCS projects, it is crucial to benchmark simulation tools to enhance confidence in their results. Inspired by a recent validation study for laboratory-scale CO$_2$ storage [@Flemisch:2024], a new comparative solution project (CSP) was launched to simulate both lab- and field-scale CO$_2$ storage [@Nordbotten:2024]. As part of the Early Access team, the authors have developed and made open the `pyopmspe11` tool which facilitates reproducible solutions to the SPE11 benchmark. This tool serves as a common starting point for developing and testing new simulation technology. Due to its user-friendly functionality (e.g., generation of different type of grids at different grid resolutions, flexibility to choose different rock and fluid properties, flexibility to define well/source locations and schedule for operations), then it is expected that its impact will extend far beyond the initial benchmark study (e.g., studies focusing on grid refinement, upscaling/coarsening approaches, numerical solvers, optimization/history matching techniques). 
 
 ![Generated model by the configuration file `spe11c_cp_ca20e6cells.txt` in the examples folder.](paper.png){ width=100% }
 
 # Statement of need
 
-Geological carbon storage (GCS) applications benefit from both commercial and open-source simulators. 
-However, using open-source simulators often involves data preprocessing and postprocessing, which can be 
-challenging for those without advanced programming knowledge. Additionally, setting up and running simulations 
-requires engineering expertise. To bridge this gap, developers can simplify the setup of numerical studies by 
-using user-friendly approaches, such as configuration files. This not only ensures result reproducibility but 
-also facilitates flexible testing of different simulator parameters and allows for easy extension to further studies.
+Geological carbon storage (GCS) applications benefit from both commercial and open-source simulators. However, using numerical simulators often involves data preprocessing and postprocessing, which can be challenging for everyone even if you know what you are doing. Additionally, setting up and running simulations requires computational expertise. To bridge this gap, developers can simplify the setup of numerical studies by using user-friendly approaches, such as configuration files. This not only ensures reproducibility of results but also facilitates flexible testing of different simulator parameters and allows for easy extension to further studies. Notable works are available in JOSS for pre-/postprocessing of simulation data, e.g., @Beucher:2019, @Sullivan:2019, @Fraters:2024, @Kaus:2024. Examples of tools which application is also in GCS and relies on the OPM Flow simulator are `pyopmnearwell` [@Landa-Marbán:2023] and `expreccs` [@Landa-Marbán:2024]. The former focuses on near well dynamics, while the latter on seamless, dynamic, and non-invasive exchange of pressure-related information between local and regional scales. 
 
[…]
