### 2024-12-03 OPM/pyopmspe11#85 @daavid00
Thanks @gassmoeller for this extensive issue.

We totally agree with you, that widely-use formats for parsing input as the ones you have mentioned should be used instead of writing own parsers. The main reason we adopted this approach was practicality, i.e., no need to define the Python variable names (i.e., less characters in the file), and only change values using the provided configuration files in the examples folder.

Then, to address this issue and comply with the guiding principles of FAIR research software you mentioned, the coming PR will add support for TOML files, i.e., we will keep the existing parsing for compatibility with previous configuration files, while TOML format will be used for input files with toml extension, e.g.,`spe11b.toml`.
