+++ docs/text/configuration_file.rst
@@ -1,6 +1,17 @@
-==================
+******************
 Configuration file
-==================
+******************
+
+In the initial development of **pyopmspe11**, the adopted configuration file format was the
+one described below, i.e., via :ref:`txt` files. The current development of **pyopmspe11** considers
+:ref:`toml` files. To keep compatibility with previous configuration files, then support for :ref:`txt` files
+will be kept, while new features will by added using :ref:`toml` configuration files.
+
+.. _txt:
+
+===
+txt 
+===
 
 The first input parameter in the configuration file is:
 
@@ -8,7 +19,7 @@ The first input parameter in the configuration file is:
     :linenos:
 
     """Set the full path to the flow executable and flags"""
-    flow --linear-solver=cprw --enable-tuning=true --enable-opm-rst-file=true --output-extra-convergence-info=steps,iterations --newton-min-iterations=1
+    flow --enable-tuning=true --enable-opm-rst-file=true --output-extra-convergence-info=steps,iterations
 
 If **flow** is not in your path, then write the full path to the executable
 (e.g., /Users/dmar/opm/build/opm-simulators/bin/flow). We also add in the same 
@@ -23,9 +34,9 @@ line as many flags as required (see the OPM Flow documentation `here <https://op
     flow simulator to improve/fix convergence and mass issues (i.e., by setting the flag \-\-linear-solver=cprw to change the linear solver,
     by tightening the cnv tolerances (\-\-tolerance-cnv), etc.).
 
-****************************
+----------------------------
 Reservoir-related parameters
-****************************
+----------------------------
 
 The following input lines in the configuration file are:
 
@@ -35,7 +46,7 @@ The following input lines in the configuration file are:
 
     """Set the model parameters"""
     spe11c master     #Name of the spe case (spe11a, spe11b, or spe11c) and OPM Flow version (master or release)
-    complete gaswater #Name of the co2 model (immiscible, convective [convective requires a Flow version newer than 22-08-2024], or complete) and co2store implementation (gaswater or gasoil [oil properties are set to water internally in OPM flow])
+    complete gaswater #Name of the co2 model (immiscible, convective, or complete) and co2store implementation (gaswater or gasoil [oil properties are set to water internally in OPM flow])
     corner-point      #Type of grid (cartesian, tensor, or corner-point)
     8400 5000 1200    #Length, width, and depth [m]
     420               #If cartesian, number of x cells [-]; otherwise, variable array of x-refinment
@@ -68,9 +79,9 @@ resulting in 18 levels. Then, the z-refinment in each of these levels is set. Se
     The top figure shows the 7 different facies, while the bottom figure shows the fipnum numbers used to identify
     the boxes (A, B, and C), sensors, and the required regions to report the benchmark data.  
 
-***********************
+-----------------------
 Soil-related parameters
-***********************
+-----------------------
 The following entries define the properties of the different facies:
 
 .. code-block:: python
@@ -108,9 +119,9 @@ The following entries define the properties of the different facies:
 
     Visualization in ResInsight of the relative permeability and capillary pressure functions in the facie 1.
 
-***********************
[…]
