+++ .github/workflows/CI.yml
@@ -12,7 +12,7 @@ jobs:
     strategy:
       fail-fast: false
       matrix:
-        python-version: ['3.8']
+        python-version: ['3.11']
         os: [ubuntu-latest]
         
     runs-on: ${{ matrix.os }}

[…]
