### 2024-12-04 OPM/pyopmspe11#83 @daavid00
Thanks @gassmoeller for your comment.

Now the [CONTRIBUTING.md](https://github.com/OPM/pyopmspe11/blob/main/CONTRIBUTING.md) includes (after the PR https://github.com/OPM/pyopmspe11/pull/97) information about:

> - Contribute to the software
> - Report issues or problems with the software
> - Seek support
>
> In addition, in the section for Contribute to the software, information of how to run the tests locally is added.

A new Section in the documentation has been also created for the Contributing guidelines (after the PR  https://github.com/OPM/pyopmspe11/pull/99):

https://opm.github.io/pyopmspe11/contributing.html
