### 2024-12-04 OPM/pyopmspe11#86 @daavid00
Many thanks @MatthewFlamm for this comment. We have added as you suggested the versioning with tags in https://github.com/OPM/pyopmspe11/pull/95. In addition, we have created the first release of the repository before starting the review process (v2024.04), and the plan is to create the next release after the reviewing process (v2024.10).
