### 2024-12-04 OPM/pyopmspe11#88 @daavid00
Thanks @MatthewFlamm for this comment, in https://github.com/OPM/pyopmspe11/pull/94 the flag **-l** is introduced to deactivate the LaTeX requirement. In addition, we have added in the documentation how to install this dependency in Linux and macOS in https://github.com/OPM/pyopmspe11/pull/99
