### 2024-12-04 OPM/pyopmspe11#89 @daavid00
We have elaborated on this at the end of the first paragraph in the Statement of need section (please see https://github.com/OPM/pyopmspe11/pull/101), thanks for the comment.
