+++ .github/workflows/CI.yml
@@ -41,11 +41,10 @@ jobs:
       run: |
         pip install --upgrade pip setuptools wheel
         pip install -r dev-requirements.txt
-        pip install opm
         
     - name: Install pyopmspe11
       run: |
-        pip install .
+        pip install .[opm]
         
     - name: Check code style and linting 
       run: |
