### 2024-12-04 OPM/pyopmspe11#82 @daavid00
Thanks @gassmoeller and @MatthewFlamm for your feedback, this PR https://github.com/OPM/pyopmspe11/pull/96 makes possible to run independently the tests, and in  https://github.com/OPM/pyopmspe11/pull/99 details on how to install the requirements to run the tests is added.
