### 2024-12-04 OPM/pyopmspe11#82 @daavid00
> This looks as if the test is trying to do the following in data.py: `float(row[0])` with `row[0] = "np.float64(0.049999999999999996)"`. I am not sure how this happens if the main workflow for the repository works correctly

Actually I also ran into this when I used Python3.12 in Ubuntu (I am a macOS user and this issue does not show up in Python3.12, and the CI used to run with Python3.8 in Ubuntu). Thanks for spotting this, and the fix has been merged also in https://github.com/OPM/pyopmspe11/pull/102, by changing from:
`str([dic["xyz"][0], dic["ymy_center"][0], dic["xyz"][2]])[1:-1]`
to
`f"{dic['xyz'][0]}, {dic['ymy_center'][0]}, {dic['xyz'][2]}"`
