### 2024-12-04 OPM/pyopmspe11#85 @daavid00
Support for toml input files has been added in https://github.com/OPM/pyopmspe11/pull/102, thanks again @gassmoeller.
