### 2024-12-04 OPM/pyopmspe11#87 @daavid00
Thanks @MatthewFlamm for this suggestion. After merging https://github.com/OPM/pyopmspe11/pull/95, which was also your feedback and very useful, then now we have extended the documentation in the Python installation to explain how one can install a specific release of this repository (https://opm.github.io/pyopmspe11/installation.html#python-package). We also mentioned that package managers such as Anaconda are not tested. Then, @totto82 and I we believe that with the provided documentation the installation of this repository could be achieved, and it is not planned to make package versions of this repository for now.
