### 2024-12-20 OPM/pyopmspe11#82 @gassmoeller
Thanks for the fix!
