### 2024-12-20 OPM/pyopmspe11#85 @gassmoeller
Thanks for the quick implementation, I am glad you picked up the idea. :tada:

> we will keep the existing parsing for compatibility with previous configuration files, while TOML format will be used for input files with toml extension, e.g.,spe11b.toml

I understand the wish to stay compatible with existing input file. I would recommend to then clearly state in the documentation that the text based format is frozen/deprecated with some message, and after a reasonable time ultimately remove the text format. Keeping two input formats of which one is increasingly outdated will split your user base and eventually confuse new users. This is something for the future of course.
