### 2024-12-20 #7357 @gassmoeller
My comments concerning the software have all been sufficiently addressed, thank you @david00! As soon as the pull request regarding the paper itself (https://github.com/OPM/pyopmspe11/pull/101) has been merged, this is ready for publication from my point of view. Thanks for pushing this along @rwegener2! And sorry it took a while to get the initial review going.
