### 2025-01-13 OPM/pyopmspe11#87 @MatthewFlamm
I don't think this issue is necessarily closed, but the changes are sufficient for purposes of https://github.com/openjournals/joss-reviews/issues/7357.
