### 2025-01-13 OPM/pyopmspe11#88 @MatthewFlamm
I haven't tested this, but having a flag to turn it off when not needed is good.
