### 2025-01-13 #7357 @MatthewFlamm
I have finished looking over the submission and repo again.  All issues/comments have been substantially addressed, and this is nearing completion from my perspective.  The last comments that need to be addressed are in https://github.com/OPM/pyopmspe11/pull/101, and then the paper will have to be quickly reviewed after it is revised.  For example, citation linkage and DOI are correct.

I think the addition of the toml format is a great outcome of this review process making this submission more closely aligned with commonly accepted data configuration practices for the community.  So kudos to @gassmoeller for suggesting it and @daavid00 for implementing it.
