### 2025-01-13 OPM/pyopmspe11#86 @MatthewFlamm
This is great.  Thanks.
