### 2025-01-15 #7357 @rwegener2
Wonderful! Many thanks again to @gassmoeller and @MatthewFlamm for your thoughtful feedback.

@daavid00 Next we will transition to the post-review phase where I will perform some checks from the editor side. Once that is complete the paper will transition to the track editor, who has the ability to fully publish the paper. I'm about to make a checklist of my own steps. I'll ping you when I have pre-publication requests on your end about dois, author checks, and similar.
