### 2025-01-15 #7357 @rwegener2
Hi @daavid00, this is a nice paper! A few points of feedback:

### General Comment
Overall, I think the paper could benefit from a bit more background knowledge on the industry benchmarks and processing workflow. For example, neither the SPE11 benchmark nor a definition for "OPM" are given. I had some confusion reading the statement of need, but I wonder if it could be solved by introducing the SPE11 benchmark in more detail. A sentence addressing the workflow could also be useful, as it would help frame the place of pyopmspe11.

### Specific Comments
- line 13: I really liked the first several sentences of your intro, the oriented me nicely. I got lost, though, when the you got to the SPE11 benchmark. It sounds like pyopmspe11 fits into a very specific part of the simulation process (pre and post processing, I'm understanding), but it's difficult to understand what that means without a bit more background on what the benchmark is. Ex. Is it a minimum uncertainty level for GCS projects? Or something more related to defining mandatory reporting values for simulations?
- line 15: "simulation technology" -- this is specifically GCS simulation technology, correct? Suggest adding "GCS"
- line 15: type -> type**s**
- line 17:  "... and schedule for operations), ~then~ it is expected that..."
- Figure 1 caption: Are there are couple high-level configurations that could be mentioned briefly in the comment? This would save a reader from needing to go look at and understand the configuration file in order to know what they are looking at
- line 22-30: somewhere in here could be a good place to set up the workflow that is typical for testing new simulation technology (ex. first define grids, then run model physics/processing, then ...)
- line 31: I'm not understanding this sentence the way it is written. Maybe "examples of pre/post processing simulation tools which also have applications in GCS and rely on the OPM flow simulator include: ..."
- line 31: Please introduce what the OPM flow simulator is and how it connects to the computational landscape of this package
- line 41: "the previous one" - does this refer a previous SPE benchmark? Or from the Class et al paper?
- line 42: who are "participants"? Are they users of the software?
- line 54: widely-use**d**
- line 55-59: this is quite a long sentence. I suggest breaking it into two sentences
- line 66: either "methods efficiently interpolate" or "method efficiently interpolates" (grammar)
- line 71: "simulation grid" and "reporting grid" are additional terms it could be helpful to define earlier on. I have a decent intuitive sense of what the simulation grid is, but I'm not sure why that is different from the reporting grid
