### 2025-01-16 #7357 @daavid00
Hi @rwegener2,

Many thanks for your comments, they really raised issues that needed to be fixed. We have addressed them in https://github.com/OPM/pyopmspe11/pull/106. Please let us know if further changes are needed before merging that PR. By implementing your comments some of the text in the paper has been relocated, please let us know if something is not clear.

Many thanks 🙂.
