### 2025-01-23 #7357 @rwegener2
> Please let us know if further changes are needed before merging that PR.

Thanks for the updates, @daavid00! I think it is a much improved paper. There were a handful of small grammatical fixes I added to the `joss` branch of the `daavid00/pyopmse11` repo. Would you be willing to allow me to push so I can share those changes with you?
