### 2025-01-23 #7357 @rwegener2
> Is it ok if I first merge https://github.com/OPM/pyopmspe11/pull/106, and after if you could please make a new PR with the grammatical fixes?

Sure thing.
