### 2025-01-24 #7357 @rwegener2
Thanks @daavid00! At this point I'll ask you to complete the tasks in section 2 of [this comment](https://github.com/openjournals/joss-reviews/issues/7357#issuecomment-2593260622). Post the dois in this issue and I'l follow up with the final set of tasks.
