### 2025-01-27 #7357 @rwegener2
Hi @daavid00, a few follow up questions:
1. Typically in JOSS submissions the zenodo archive has a title which matches the paper title. Is there a reason you prefer this title to a title matching the paper?
2. The same is also true for author lists between the archive and the paper. Was there intention in choosing a different author list for the zenodo archive?
3. I see you shared two archive DOIs, one on Zenodo and one on Figshare. Do you have a preference which one we link to the JOSS paper?
