### 2025-01-27 #7357 @daavid00
Hi @rwegener2,

For zenodo I just made the release without changing the standard settings, that is why the title does not match the paper and also contributors that commits were merged for this release also appears, but not @totto82 (this is strange since for JOSS @totto82 commit contributions are shown, but not in Zenodo). Then, if you could please use the DOI on Figshare to link the paper, that would be great.
