### 2025-01-27 #7357 @rwegener2
> Then, if you could please use the DOI on Figshare to link the paper, that would be great.

The figshare DOI works great for us. We usually prefer the titles and author lists match the papers and that appears to be the case with the figshare archive.
