### 2025-01-27 #7357 @danielskatz
The error message says there are some ISBN fields in the bib file that aren't actually ISBNs, as they have 9 characters, and ISBNs have to have at least 10. I didn't look at the file to see the specific problems, but I hope this lets you find them.
