### 2025-01-27 #7357 @daavid00
Hi @rwegener2, it seems is due to the isbns from https://github.com/OPM/pyopmspe11/blob/main/paper/paper.bib#L38 and https://github.com/OPM/pyopmspe11/blob/main/paper/paper.bib#L139 which have a length of 9. I could remove these from the .bib.
