### 2025-01-27 #7357 @danielskatz
See https://en.wikipedia.org/wiki/ISBN for some info on ISBNs...

But just removing them unless you are sure they are correct is probably best, and if they are correct SBNs from before 1966 (though I think this is rare), then should have a leading 0 added to convert them to ISBNs.

Alternatively, you could try to find the correct ISBNs
