### 2025-01-27 #7357 @danielskatz
I think the numbers in these fields are actually ISSNs - https://en.wikipedia.org/wiki/ISSN
