### 2025-01-27 #7357 @daavid00
Thanks @danielskatz for the information, you are right, they are ISSNs, it seems there is a bug in the Springer exporting bibliography system. I will update the bibliography then.
