+++ paper/paper.bib
@@ -35,7 +35,7 @@ date-added = {2024-06-25 15:36:09 +0200},
 date-modified = {2024-06-25 15:36:09 +0200},
 doi = {10.1007/s11242-023-01977-7},
 id = {Flemisch2024},
-isbn = {1573-1634},
+issn = {1573-1634},
 journal = {Transport in Porous Media},
 number = {5},
 pages = {865--912},
@@ -136,7 +136,7 @@ date-added = {2024-12-03 12:41:48 +0100},
 date-modified = {2024-12-03 12:41:48 +0100},
 doi = {10.1007/s10596-009-9146-x},
 id = {Class2009},
-isbn = {1573-1499},
+issn = {1573-1499},
 journal = {Computational Geosciences},
 number = {4},
 pages = {409--434},
