### 2025-01-28 #7357 @daavid00
Hi @rwegener2 ,

The ISSNs have been updated in https://github.com/OPM/pyopmspe11/pull/108, should I make a new release (i.e., deleting the current one) in figshare and share the doi?
