### 2025-01-28 #7357 @rwegener2
> The ISSNs have been updated in https://github.com/OPM/pyopmspe11/pull/108, should I make a new release (i.e., deleting the current one) in figshare and share the doi?

Thanks, @daavid00! Since only the paper was changed you shouldn't need to create a new upload and DOI unless you'd like to.
