### 2025-01-28 #7357 @editorialbot
:wave: @openjournals/ese-eics, this paper is ready to be accepted and published.
