### 2025-01-28 #7357 @rwegener2
@daavid00 The paper looks good!

To follow up again on your question, JOSS does not require that you create a new release and version if only the paper was changed. If, however, you would prefer that everything connected to the JOSS publication matches then you can create a new release and archive copy.

If you'd like to make a new archive DOI and version please:
1. comment here to let me know you'd like to to that (if it will be more than a few hours before you can complete it)
2. comment with new version and archive DOI

Then I'll **recommend-accept** again.

Please still let me know if you're okay with the version and DOI as is, so the track editor Kristen knows if she can come complete the publication.
