### 2025-01-28 #7357 @daavid00
The current version and DOI are OK.

Many thanks for all your help @rwegener2 🙂.
