### 2025-01-29 #7357 @daavid00
Hi @rwegener2,
I have checked the paper PDF including the one in the deposit XML and they look good. Should I run the command `@editorialbot accept` or that should be run by you?
Many thanks again.
