### 2025-01-30 #7357 @kthyng
@daavid00 I'm not so used to figshare so I might be missing it, but I don't see the version associated with your JOSS review reflected in your software archive. Can you verify that it is there, and if it isn't, can you find a good place to include it?
