### 2025-01-30 #7357 @daavid00
Hi @kthyng,

Thanks for taking over this contribution. I have updated the description in figshare with the version (v2024.10) and also added the link to the joss review process and acknowledgment to the reviewers and editor (this is also my first time with figshare but so far it seems quite user friendly). Please let me know if something needs to be updated.
