### 2025-01-30 #7357 @daavid00
Many thanks @gassmoeller, @MatthewFlamm, @rwegener2, and @kthyng for your time and excellent work in this contribution. I have just signed in as a reviewer, looking forward to being part of the JOSS community 🙂.
